% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_GPL.R
\name{read_GPL}
\alias{read_GPL}
\title{Read and Process GEO Platform Annotation}
\usage{
read_GPL(
  f_soft,
  drop_na = TRUE,
  save_ids_only = TRUE,
  save_rda = NULL,
  rda_name = NULL
)
}
\arguments{
\item{f_soft}{Character. Path to the GEO platform annotation file in SOFT format.}

\item{drop_na}{Logical. Whether to drop rows with missing gene symbols. Default is \code{TRUE}.}

\item{save_ids_only}{Logical. If \code{TRUE}, only the processed probe-to-gene mapping is returned.
Otherwise, additional raw and intermediate data are included in the output. Default is \code{TRUE}.}

\item{save_rda}{Character. Directory where the \code{.rda} file should be saved. Default is \code{NULL}, meaning no file will be saved.}

\item{rda_name}{Character. Name of the \code{.rda} file (without extension). If \code{NULL}, it defaults to the platform ID with a \verb{_bioc} suffix.}
}
\value{
If \code{save_ids_only} is \code{TRUE}, a data frame with probe IDs and their associated gene symbols.
Otherwise, a list containing:
\describe{
\item{ids}{Data frame. Processed probe-to-gene mappings.}
\item{gpl_ori}{Data frame. Original platform annotation table.}
\item{gpl_anno}{Data frame. Intermediate processed annotation table.}
}
}
\description{
This function reads a GEO platform (GPL) annotation file in SOFT format, processes
probe-to-gene mappings, and optionally saves the result in \code{.rda} format.
}
\details{
The function parses the platform annotation table using \code{GEOquery::parseGEO} and extracts relevant columns for
probe-to-gene mappings. It applies further filtering and formatting to clean the data, ensuring consistency.

If \code{save_rda} is specified, the processed data is saved as an \code{.rda} file in the given directory, mimicking
the storage format used in the \code{AnnoProbe} package.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
f_soft <- "./inst/extdata/GSE106172/soft/GSE106172_family.soft.gz"
ids <- read_GPL(f_soft = f_soft, drop_na = TRUE, save_ids_only = TRUE)
full_data <- read_GPL(f_soft = f_soft, save_ids_only = FALSE)
\dontshow{\}) # examplesIf}
}
